WIP: Refactor to enable integration via JPY
